refactor: split commands into separate files
