+++ .changeset/split-commands.md
@@ -1,0 +1,5 @@
+---
+"@aliou/pi-processes": patch
+---
+
+Split commands into separate files for better organization
