+++ src/commands/ps-clear-command.ts
@@ -1,0 +1,14 @@
+import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
+import type { ProcessManager } from "../manager";
+
+export function registerPsClearCommand(
+  pi: ExtensionAPI,
+  manager: ProcessManager,
+): void {
+  pi.registerCommand("ps:clear", {
+    description: "Clear finished processes",
+    handler: async (_args, _ctx) => {
+      manager.clearFinished();
+    },
+  });
+}
